ARM: imx6: Map the GIC CPU registers as MT_DEVICE_NS
Statically map the SoCFPGA's memory space at PA 0x00a00000 +1MiB to 0xff000000
and set type of this VA to be MT_DEVICE_NS. This PA contains the i.MX6's GIC
CPU registers at offset +0x100. This area is right past VMALLOC_END and well
below the optional start of per-CPU highmem entries. We can thus use this area
to place our mapping here and be sure it's not in anyone's way.

All accesses to VA 0xff000000 generate non-secure accesses on the bus and we
can leverage this property to generate non-secure read of the GIC INTACK (IAR)
register. This non-secure read will never return an FIQ interrupt number.

Signed-off-by: Marek Vasut <marex@denx.de>