gic: Mask interrupt before reconfiguring it
To prevent generation of interrupts with undefined properties, mask the
interrupts before changing their PRIority and GROUP , that is, before
switching them between FIQ and IRQ mode. The problem here is that the
change of interrupts' PRIority and GROUP properties is not an atomic
operation, but is instead done via two separate writes. Therefore, if
the interrupt happened inbetween these two writes, an requests with
undefined properties would be generated.

Moreover, remove any possible dangling interrupts with the old properties
from the GIC's pending interrupts stack. This is necessary again to make
sure we won't be getting any spurious interrupts.

Signed-off-by: Marek Vasut <marex@denx.de>